+++ SeaBattleTrophy/Maths/Geometry/Vector2D.cs
@@ -4,21 +4,21 @@ namespace Maths.Geometry
 {
     public struct Vector2D : IEquatable<Vector2D>
     {
-        public float X { get; }
-        public float Y { get; }
+        public double X { get; }
+        public double Y { get; }
 
-        public Vector2D(float x, float y)
+        public Vector2D(double x, double y)
         {
             X = x;
             Y = y;
         }
 
-        public static Vector2D operator *(Vector2D vector, float val)
+        public static Vector2D operator *(Vector2D vector, double val)
         {
             return new Vector2D(vector.X * val, vector.Y * val);
         }
 
-        public static Vector2D operator /(Vector2D vector, float val)
+        public static Vector2D operator /(Vector2D vector, double val)
         {
             return new Vector2D(vector.X / val, vector.Y / val);
         }
@@ -28,28 +28,28 @@ namespace Maths.Geometry
             return new Vector2D(vector0.X + vector1.X, vector0.Y + vector1.Y);
         }
 
-        public float Dot(Vector2D vector)
+        public double Dot(Vector2D vector)
         {
             return X * vector.X + Y * vector.Y;
         }
 
-        public float SquaredLength()
+        public double SquaredLength()
         {
             return X * X + Y * Y;
         }
 
-        public float Length()
+        public double Length()
         {
-            return (float)Math.Sqrt(SquaredLength());
+            return Math.Sqrt(SquaredLength());
         }
 
-        public Vector2D Rotate(float degrees)
+        public Vector2D Rotate(double degrees)
         {
             var radian = degrees / 180 * Math.PI;
             var x = X * Math.Cos(radian) - Y * Math.Sin(radian);
             var y = X * Math.Sin(radian) + Y * Math.Cos(radian);
 
-            return new Vector2D((float)x, (float)y);
+            return new Vector2D(x, y);
         }
 
         public bool Equals(Vector2D other)
@@ -57,9 +57,9 @@ namespace Maths.Geometry
             return X.Equals(other.X) && Y.Equals(other.Y);
         }
 
-        public static Vector2D DirectionFromAngle(float angle)
+        public static Vector2D DirectionFromAngle(double angle)
         {
-            return new Vector2D((float)-Math.Sin(angle * Math.PI / 180.0), (float)Math.Cos(angle * Math.PI / 180.0));
+            return new Vector2D(-Math.Sin(angle * Math.PI / 180.0), Math.Cos(angle * Math.PI / 180.0));
         }
     }
 }
